feat(minimumPluginVersion): implement minimum plugin version
In order to use certain features in testsuites that require a specific plugin version, add an attribute to define a minimum version of the plugin in order for the testsuite to work.

* add the minimumVersion to Testsuite.Skip to use the existing skip-functionality
* move the plugin version from plugin.yaml to version.go in order to be able to access it when using just the binary untt
* write the plugin version from version.go to plugin.yaml during build/dist/etc.
* add tests and update snapshots to consider new number of tests and templates in test/data/v3/basic

defines an optional minimumVersion attribute on the testsuite and checks it against the plugin version defined in plugin.yaml. fails tests that require a higher version of the plugin.